display the version at /version
