+++ src/Directory/Biobanks.Web/HtmlHelpers/VersionHelper.cs
@@ -9,28 +9,18 @@ namespace Biobanks.Web.HtmlHelpers
     {
         public static string AssemblyVersion(this HtmlHelper helper)
         {
-            return GetAspNetAssembly().GetName().Version.ToString();
+            return Utilities.VersionHelper.GetAspNetAssembly().GetName().Version.ToString();
         }
 
         public static string InformationalVersion(this HtmlHelper helper)
         {
             //InformationalVersion can contain strings, not just numbers
-            //so is useful for Semantic Versioning http://semver.org
+            //so is useful for Semantic Versioning http://semver.org            
 
-            //http://stackoverflow.com/questions/7770068/get-the-net-assemblys-assemblyinformationalversion-value
-            var attr = Attribute.GetCustomAttribute(
-                    GetAspNetAssembly(),
-                    typeof(AssemblyInformationalVersionAttribute))
-                as AssemblyInformationalVersionAttribute;
+            var attr = Utilities.VersionHelper.GetVersionNumber();            
 
             //return the assembly version if no informational version
-            return attr?.InformationalVersion ?? AssemblyVersion(helper);
-        }
-
-        private static Assembly GetAspNetAssembly()
-        {
-            //http://stackoverflow.com/questions/4277692/getentryassembly-for-web-applications
-            return HttpContext.Current.ApplicationInstance.GetType().BaseType?.Assembly;
-        }
+            return attr ?? AssemblyVersion(helper);
+        }        
     }
 }
