+++ src/Directory/Biobanks.Web/Utilities/VersionHelper.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Web;
+
+namespace Biobanks.Web.Utilities
+{
+    public class VersionHelper
+    {
+        public static string GetVersionNumber()
+        {
+            //http://stackoverflow.com/questions/7770068/get-the-net-assemblys-assemblyinformationalversion-value
+            var attr = Attribute.GetCustomAttribute(
+                    GetAspNetAssembly(),
+                    typeof(AssemblyInformationalVersionAttribute))
+                as AssemblyInformationalVersionAttribute;
+
+            return attr?.InformationalVersion;
+        }
+
+        public static Assembly GetAspNetAssembly()
+        {
+            //http://stackoverflow.com/questions/4277692/getentryassembly-for-web-applications
+            return HttpContext.Current.ApplicationInstance.GetType().BaseType?.Assembly;
+        }
+    }
+}
